revised Java

+++ transfer.java
@@ -1,8 +1,8 @@
 public class transfer {
     public static void main(String[] args) {
-        long currentBalance =  2_000_000_000;
-        long transferAmount = 500_000_000;
-        long total;
+        int currentBalance =  2_000_000_000;
+        int transferAmount = 500_000_000;
+        int total;
         total=currentBalance+transferAmount;
 
         System.out.println(total);
